Bug: safeTransfer when childTransfer in composable
When the caller of safetransfer() is a contract, that would be the composable when it is trying to move ownership of some token. In material in data we put two inputs in bytes what normally should be one because of dependencies problems of other contracts. On that way is caller direct the wallet owner, but when the caller is other contract the input of data is only token ID so the function that separate the token ID from the amount will corrupt the input. As a solution, we did put statement that says: when the caller is a contract, just safeTranfer like what it normally is